Update patch set 2
Patch Set 2:

The digest does get printed on a single line in dmesg output for boot, but logcat doesn't like it.
Anyway, we can drop it but this points out another problem with the current logging callback in the SELinux jni code not preserving the type/priority info (in which case it would have been info-level, not error).  I'll fix that separately.

Patch-set: 2