Update patch set 1
Patch Set 1:

I see this new message in logcat with this change:

E/SELinux ( 1518): SELinux:  Could not open /data/system/packages.list:  Permission denied.

Process 1518 is android.process.media, which lacks DAC permissions to read that file.  libselinux only tries to open the file if the caller is calling restorecon on a /data/data or /data/user/N pathname.  I assume this is for /data/data/com.android.providers.downloads/cache download_file labeling?  The error is harmless as libselinux will then just not find any matching pkgInfo and will fallback to just using file_contexts as before, but it does show up in the logcat output.  Not sure if we should drop the log statement to avoid confusion or not.

Patch-set: 1