Update patch set 1
Patch Set 1:

At the moment this doesn't cause any obvious issues because com.android.providers.downloads/cache is restorecon'd by DownloadProvider each time and since it lacks read permission to packages.list, it cannot find the package info and therefore cannot perform the package-based lookup.  So it is still labeled download_file in spite of the boot restorecon. Looks like the com.android.settings/files/wallpaper entry is a leftover of older Android and obsoleted by the /data/system/users/N/wallpaper files, which are not affected by this change since they are not in package directories.

However, the cache dir is only labeled correctly as a side effect of not being able to read packages.list, so we do not want to rely on that.

Unclear how to distinguish the case.

Patch-set: 1