Update patch set 1
Patch Set 1:

We had in the past considered extending seapp_contexts to support different labels on specific subdirs of the package directories, so we could introduce something like that for the cache subdir labeling so that restorecon() would always be consistent.  A new name=value pair, e.g. subdir=.

Alternatively, we could create an extended version of selabel_lookup() to return the full info for the matching entry and have restorecon decide whether to override the file_contexts match with the seapp_contexts match based on the specificity of the match (e.g. if the matching regex in file_contexts was only /data/data(/.*)? or /data(/.*)?, then let seapp_contexts override; if it is a path to a specific package directory, then let file_contexts take precedence).

Patch-set: 1