Update patch set 1
Patch Set 1: Code-Review+1

This is fine to me.

For robustness, should we also modify file_contexts so that /data/data explicitly never has a label.

Instead of:

# App sandboxes
/data/data/.*       u:object_r:app_data_file:s0

we have

# App sandboxes
/data/data       u:object_r:app_data_file:s0
/data/data.*     <<none>>

We could also do the same thing for /data/users/

It almost feels like we could substitute this patch with a change to the rules in file_context, and have essentially the same effect.

I don't know if this is a good idea or not. Just wanted to throw it out there as one possibility. Let me know your thoughts.

Patch-set: 1
Label: Code-Review=+1