Disable restorecon_last usage for restorecon /data/data.
Since /data/data and /data/user labeling is based on seapp_contexts and seinfo
assignments rather than file_contexts, we do not want to get or set
the security.restorecon_last attribute on these directories as that will
incorrectly skip subsequent recursive restorecon calls on these directories.
Also, setting a security. attribute other than selinux or capability
requires CAP_SYS_ADMIN and thus attempting to do this would trigger
denials for installd, which handles relabeling of these directories.

Ensuring that /data/data and /data/user are only relabeled when
necessary is the responsibility of the caller of restorecon, in
this case the system_server and installd.

Change-Id: I33476236c493f9749a3da068afd83d9f279409a9
Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>