Update patch set 7
Patch Set 7:

So I think this is basically ready to go aside from updating SELinuxPolicyInstallReceiver to include one of these selinux_version files (and mac_permissions.xml) in its policy bundle format and the files created under /data/security.  But the question is whether this mechanism suffices to address prior concerns or not.

Patch-set: 7