Don't set restorecon_last on all subdirectories
When restorecon_recursive is called, we set the directory xattr
"security.restorecon_last" to the hash of /file_contexts.
This allows us to do automatic relabeling when /file_contexts
changes.

Prior to this change, we were also setting the xattr for all
subdirectories of the directory. Doing so is unnecessary because
we never look at the value.

Remove setting the xattr for subdirectories, but continue to set
the xattr for the directory itself.

Change-Id: Id81d1e24209e195c559b4e382bee42ddd48a7593