Update patch set 1
Patch Set 1:

(2 comments)

This may have been discussed before, but moving selinux_android_prop_context_handle() seems like the right thing to do.  In the process of doing that, I toyed with some asthetic changes in android.c, upon which we may want to expand if this is deemed the direction in which we'd like to go.

Patch-set: 1
Label: Verified=0