Update patch set 1
Patch Set 1:

> Is their some plan to drop the dup static field set to 0 in the
 > code, why even have that? Lets just code it to reject dupes
 > instantly.

I'm trying to minimize the code change and impact for what should be a very unusual case.  Even if you had dups in your source seapp_contexts files, checkseapp was already deduplicating those (last one wins) prior to my recent change, so it is unlikely that we'll ever see this happen in a device seapp_contexts.  Since we already qsort the array, it was simplest to just put a test in the compare callback for the qsort function, but that gives me no way to pass back information except via a global.  So I just did that.  And, no, I wasn't planning on changing it.  The seapp_contexts array is already global so this isn't a fundamental change.

Patch-set: 1
Reviewer: Gerrit User 1010111 <1010111@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0