libselinux:  Reject duplicate entries within seapp_contexts.
Presently it will permit duplicate entries (either duplicated
on input selectors or entirely) and whichever one ended up first
in the sorted list would win (and this would be arbitrary as
the comparison function would return 0 since they have the same
input selectors).

Treat it as an error instead and fail.

Change-Id: I611515699b35b04dfc5c5020b92a88bff24ca606
Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>