Update patch set 1
Patch Set 1:

If qsort_r is our only concern, we can expose this in bionic. The implementation is already checked in, and we have an open bug (bug 17203231) to add this functionality to bionic.

The version checked into bionic is the BSD version, which differs from the GNU version in the order of the arguments.

Patch-set: 1