Update patch set 1
Patch Set 1:

(1 comment)

I'm assuming the bug was on the memcmp as well as using a global index rather than per file, causing subseuent opens to fail on non-existent files, is that right?

The index idiom seems a bit wonky compared to how it was done (all or nothing approach), which gave you a bit more flexibility on load sources. But I really don't have a strong opinion on direction as I am unaware of how people are actually using /data reload support presently.

Patch-set: 1