Update patch set 1
Patch Set 1:

(1 comment)

This isn't fixing a bug per se, just a lack of robustness in the face of an incomplete set of files under /data/security/current.  Originally at each point where we open a policy file, we tried first to open the /data/security file and then fell back to the / file if that failed.  That however was disabled and then replaced with the new logic in order to address the problem of ensuring that after an OTA, Android would not keep trying to use an old policy update.  At the time, IIRC, the view was that we want to ensure consistency among the policy files, so we didn't want to risk mixing a /data/security/current/sepolicy with a /file_contexts or vice versa, which led to the current logic.  But the current logic isn't very robust in the presence of missing files, so I just wanted to improve that situation.  For the primary intended use case (policy updates via the SELinuxPolicyInstallReceiver), we should always have a complete set of files so it would be a bug to be missing a file.  For the dontaudit case, it is a bit less convenient, but that is a developer-only work flow, so I am unclear on whether it is sufficient motivation to support partial sets of files.

Patch-set: 1
Label: Verified=0