Update patch set 1
Patch Set 1: Code-Review+1

toolbox restorecon requires -D to restorecon files under /data/data.  By default, this is disabled because it is the job of the system_server/installd and depends on package metadata.  It is actually the system_server and installd that performs the restorecon of package directories on upgrades; the init.rc restorecon also does not descend into /data/data.
I was confused however that you said you had this issue on upgrade to L because I don't think L included the pcre support. We however did back-port it to our seandroid-5.* branches, so if using those, you would encounter this.

Patch-set: 1
Reviewer: Gerrit User 1010111 <1010111@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1