Update patch set 1
Patch Set 1:

Agreed, but the intent is to _not_ add libpcre and others in the first wave of adjustments to DISTTOOLs. I decided to draw the line at libcutils, liblog and libselinux in order to unblock the addition of the new extensible android_filesystem_config.h feature. Coems down to 'when does this stop' ;-> and it ends up being a complete wipe of STATIC libraries ;-}

pcre needs a section for host shared library ...

Patch-set: 1