Update patch set 1
Patch Set 1:

Good to know about the whitelist, for now we will _not_ do zipalign for SDK and SDK+windows, and merely acknowledge that we have removed one of the blocking reasons for a conversion from static to shared library.

ToDo: Followup later with the SDK team to inform them, so they can add libselinux.dll/libselinux.so to the deliverables.

Patch-set: 1