Update patch set 1
Patch Set 1:

It changes the effect of restorecon /path/to/symlink (only).
Looking at upstream restorecon (policycoreutils/setfiles/restore.c), if the target is a symlink, it computes the realpath of the prefix, re-appends the symlink name, relabels the symlink if necessary, and then relabels the target if it exists too.

Patch-set: 1