Update patch set 2
Patch Set 2:

Actually, I think misunderstood the architecture. It looks like the validate callback is called on the open and thus I have no clue whats being processed. I think I understand what your saying a but more clearly now. It seems as though some of this needs to live in libselinux, but only for validation not the other paths.

Patch-set: 2