Update patch set 2
Patch Set 2:

It seems as though if liblog and libpackageparser are not built into libselinux as a WHOLE_STATIC archive, that the linking starts to fail when building init.

If you look at the next change where I add libpackageparse dependency, I add it as a WHOLE_STATIC, if I don't have this as WHOLE_STATIC then I get linking issues on init.

Patch-set: 2
Label: Verified=0