Update patch set 2
Patch Set 2:

Unfortunately LOCAL_STATIC_LIBRARIES isn't transitive in the current build system. That means if you have "LOCAL_STATIC_LIBRARIES := liblog" in the module definition of libselinux.a, any thing that links libselinux.a won't get liblog.a automatically. You have to add liblog.a to the executable/shared library that links libselinux.a.

Patch-set: 2