Update patch set 1
Patch Set 1:

This one isn't strictly required but addresses an issue I noticed when I first updated the SELinuxHostTest.  It was creating the copy of the device file_contexts.bin file with a .tmp suffix, and this was then preventing selabel_open() from recognizing it as a .bin file and therefore trying to parse it as a text file.  I already adjusted SELinuxHostTest to create it with a .bin suffix, so it works as is, but I thought it would be less fragile to have it probe by magic number instead.  load_mmap() was already checking the magic number; we just weren't using that to automatically recognize whether the original path was a .bin file.

Patch-set: 1