Update patch set 1
Patch Set 1:

NB This will trigger a restorecon_recursive of /data and other partitions on upgrade, regardless of whether or not file_contexts content has truly changed, simply because we are now hashing file_contexts.bin and therefore getting a different result.  But I don't think that's a big deal as it is unlikely that the next release won't also have some change in file_contexts content that would have triggered it.

Patch-set: 1