Update patch set 1
Patch Set 1:

Do we plan on having an Android specific fork of libselinux indefinitley? It would be nice to see if this can be done in a way where we can skip processing but not fork us further off of mainline.

As to what that change would look like, I am not all too sure. Perhaps SDS can chime in.

Patch-set: 1