Update patch set 6
Patch Set 6:

There are tests in CTS that verify that a number of processes are in their proper domains. Probably not sufficient.

There has been a lot of work on merging with upstream selinux so we can take advantage of the testing there.

Patch-set: 6
Reviewer: Gerrit User 1054468 <1054468@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0