Update patch set 1
Patch Set 1:

This change breaks the ability to restorecon by symlink name in init.<board>.rc files, previously relied upon to label e.g. /sys/class/leds/red/brightness without requiring unique file_contexts entries for every board. See:
http://marc.info/?l=seandroid-list&m=144170540914745&w=2

Admittedly, that was a hack, but it did work in Android 5.0, so it is a bit of a regression.

As I have no visibility into the bug, I don't know the original motivation for this change and what might be relying upon the new behavior.  The change makes sense for typical usage, but it doesn't offer any way to use the original behavior when that is actually desired.

Patch-set: 1