Update patch set 1
Patch Set 1:

So this would allow labeling a symlink file but not a symlink directory, is that correct? What if I want to follow the symlink file, do I have to get the realpath and add that to the fc entries? I dunno, this seems odd to me. I can never recall all the modes in file_contexts but could we use that to determine what they want? Perhaps extend it if it doesn't fit our needs currently?

Patch-set: 1