Update patch set 1
Patch Set 1:

I'm not sure what you mean by a symlink directory.  The original change: https://android-review.googlesource.com/#/c/154756/ went in to prevent restorecons of relative paths or symlinks from being used to relabel the pointed-to files with an improper label.  

That change, however, always followed symlinks in creating the realpath, which meant that a direct restorecon of a symlink path could not relabel the link file itself.  This corrects that.  The underlying issue of how we should treat symlink labeling in general is worth visiting, but the general preference is to use real paths, as evidenced by the previous patch. It should be noted, though, that the original patch still has the issue of following symlinks to potentially "reset" a file label to that in file_contexts and also does nothing about different rules for two hard-links.

Patch-set: 1
Label: Verified=0