Update patch set 3
Patch Set 3:

Have requested access to be made public.  In the meantime, the crux of the argument for this change is that the getpidcon(0, ) getcon() equivalence seems to be an implementation convenience rather than documented anywhere and that calling a getpidcon for pid 0 actually breaks the documentation that does exist: it does not get the context for the specified pid.

Patch-set: 3