Update patch set 3
Patch Set 3:

I agree that this could be interpreted as improper input validation, especially given precedent for other pid-based calls.  This is only an issue for us due to binder's one-way implementation, which I'd like to replace with passing context over binder anyway. I'll defer to nnk@ as to whether or not I should take this upstream given the high probability of its rejection.  As always, thanks for the input.

Patch-set: 3
Label: Verified=0