Support loading sepolicy file from /odm or /vendor partition.
This change allows ODM or VENDOR to put sepolicy files into their
own partition after extending Google's general sepolicy files.

In practice, the odm sepolicy is a superset of the vendor one, and the
vendor sepolicy is a superset of the root one. First one we find gets
loaded, in the order odm -> vendor -> root.

Bug: 25883251
Change-Id: Ibcde6211e92944aaf75ed4fb1875fb559f8f32fe