Update patch set 3
Patch Set 3:

FYI: the internal buffers for passwd_state_t and group_state_t are limited currently to 32 characters for the pw_name and gr_name respectively. We can enforce this in the build tools that create entries in /system/etc/passwd and /system/etc/group respectively, or in the parser code that reads them. We need to know if there is a limit for /data/system/packages.list though, and maybe we should settle on that limit instead?

Do we want to consider adding something like __GETPW_R_SIZE_MAX in pwd.h and __GETGR_R_SIZE_MAX in grp.h before we consider any adjustments that utilize getpw*_r and getgr*_r functions?

enh@, what is your feeling about this?

Patch-set: 3