Update patch set 2
Patch Set 2:

> > (1 comment)
 > 
 > Can you see what this looks like with an ASANified build?

This version is fine in its exact usage, but as Janis points out if someone calls regex_version() multiple times on PCRE1 they get N strings, on PCRE2, they get 1 string, which is ripe for double free.

I think the simplest form is the newest patch uploaded.

Whats the plan for up-streaming the PCRE2 changes? Ideally we want to eventually coalesce the upstream and Android forks.

Patch-set: 2