fix lax service context lookup
We use the same lookup function for service contexts
that we use for property contexts. However, property
contexts are namespace based and only compare the
prefix. This may lead to service associations with
a wrong label.

This patch introduces a stricter lookup function for
services contexts. Now the service name must match
the key of the service label exactly.

Test: bullhead builds and boots

Bug: 31353148
Change-Id: I16dd276e1e74c0d04af4e067992385be5499b01d