Update patch set 4
Patch Set 4:

> bump Janis, do we still need this?

Well it does not do harm either way. In the way we use libselinux, this bug does not present a problem to system health. And when the android libselinux fork will finally be killed - thanks to your effort - this patch will become obsolete because upstream does not exhibit this bug.

Patch-set: 4